Move custom ingredient code to new Recipe API module
