+++ fabric-data-generation-api-v1/build.gradle
@@ -6,8 +6,8 @@ moduleDependencies(project, [
 		'fabric-registry-sync-v0',
 		'fabric-networking-api-v1',
 		'fabric-resource-conditions-api-v1',
-		'fabric-ingredient-api-v1',
-		'fabric-item-group-api-v1'
+		'fabric-item-group-api-v1',
+		'fabric-recipe-api-v1',
 ])
 
 dependencies {
