--- fabric-ingredient-api-v1/build.gradle
@@ -1,10 +1,0 @@
-archivesBaseName = "fabric-ingredient-api-v1"
-version = getSubprojectVersion(project)
-
-loom {
-	accessWidenerPath = file('src/main/resources/fabric-ingredient-api-v1.accesswidener')
-}
-
-moduleDependencies(project, [
-		"fabric-networking-api-v1",
-])
