--- fabric-ingredient-api-v1/src/client/java/net/fabricmc/fabric/impl/ingredient/client/CustomIngredientSyncClient.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.impl.ingredient.client;
-
-import java.util.concurrent.CompletableFuture;
-
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.networking.v1.ClientLoginNetworking;
-import net.fabricmc.fabric.impl.ingredient.CustomIngredientSync;
-
-/**
- * @see CustomIngredientSync
- */
-public class CustomIngredientSyncClient implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-		ClientLoginNetworking.registerGlobalReceiver(CustomIngredientSync.PACKET_ID, (client, handler, buf, listenerAdder) -> {
-			if (buf.readVarInt() != CustomIngredientSync.PROTOCOL_VERSION) {
-				// Unknown protocol version ?
-				return CompletableFuture.completedFuture(null);
-			}
-
-			return CompletableFuture.completedFuture(CustomIngredientSync.createPacket());
-		});
-	}
-}
