--- fabric-ingredient-api-v1/src/main/java/net/fabricmc/fabric/impl/ingredient/CustomIngredientImpl.java
@@ -1,143 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.impl.ingredient;
-
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Stream;
-
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import org.jetbrains.annotations.Nullable;
-
-import net.minecraft.item.ItemStack;
-import net.minecraft.network.PacketByteBuf;
-import net.minecraft.recipe.Ingredient;
-import net.minecraft.util.Identifier;
-
-import net.fabricmc.fabric.api.ingredient.v1.CustomIngredient;
-import net.fabricmc.fabric.api.ingredient.v1.CustomIngredientSerializer;
-import net.fabricmc.fabric.impl.ingredient.builtin.AndIngredient;
-import net.fabricmc.fabric.impl.ingredient.builtin.DifferenceIngredient;
-import net.fabricmc.fabric.impl.ingredient.builtin.NbtIngredient;
-import net.fabricmc.fabric.impl.ingredient.builtin.OrIngredient;
-
-public class CustomIngredientImpl extends Ingredient {
-	// Static helpers used by the API
-
-	public static final String TYPE_KEY = "fabric:type";
-	public static final int PACKET_MARKER = -1;
-
-	static final Map<Identifier, CustomIngredientSerializer<?>> REGISTERED_SERIALIZERS = new ConcurrentHashMap<>();
-
-	public static void registerSerializer(CustomIngredientSerializer<?> serializer) {
-		Objects.requireNonNull(serializer.getIdentifier(), "CustomIngredientSerializer identifier may not be null.");
-
-		if (REGISTERED_SERIALIZERS.putIfAbsent(serializer.getIdentifier(), serializer) != null) {
-			throw new IllegalArgumentException("CustomIngredientSerializer with identifier " + serializer.getIdentifier() + " already registered.");
-		}
-	}
-
-	@Nullable
-	public static CustomIngredientSerializer<?> getSerializer(Identifier identifier) {
-		Objects.requireNonNull(identifier, "Identifier may not be null.");
-
-		return REGISTERED_SERIALIZERS.get(identifier);
-	}
-
-	static {
-		CustomIngredientSerializer.register(AndIngredient.SERIALIZER);
-		CustomIngredientSerializer.register(OrIngredient.SERIALIZER);
-		CustomIngredientSerializer.register(DifferenceIngredient.SERIALIZER);
-		CustomIngredientSerializer.register(NbtIngredient.SERIALIZER);
-	}
-
-	// Actual custom ingredient logic
-
-	private final CustomIngredient customIngredient;
-
-	public CustomIngredientImpl(CustomIngredient customIngredient) {
-		super(Stream.empty());
-
-		this.customIngredient = customIngredient;
-	}
-
-	@Override
-	public CustomIngredient getCustomIngredient() {
-		return customIngredient;
-	}
-
-	@Override
-	public boolean requiresTesting() {
-		return customIngredient.requiresTesting();
-	}
-
-	@Override
-	public ItemStack[] getMatchingStacks() {
-		if (this.matchingStacks == null) {
-			this.matchingStacks = customIngredient.getMatchingStacks();
-		}
-
-		return this.matchingStacks;
-	}
-
-	@Override
-	public boolean test(@Nullable ItemStack stack) {
-		return stack != null && customIngredient.test(stack);
-	}
-
-	@Override
-	public void write(PacketByteBuf buf) {
-		// Can be null if we're not writing a packet from the PacketEncoder; in that case, always write the full ingredient.
-		// Chances are this is a mod's doing and the client has the Ingredient API with the relevant ingredients.
-		Set<Identifier> supportedIngredients = CustomIngredientSync.CURRENT_SUPPORTED_INGREDIENTS.get();
-
-		if (supportedIngredients != null && !supportedIngredients.contains(customIngredient.getSerializer().getIdentifier())) {
-			// The client doesn't support this custom ingredient, so we send the matching stacks as a regular ingredient.
-			// Conveniently, this is exactly what the super call does.
-			super.write(buf);
-		} else {
-			// The client supports this custom ingredient, so we send it as a custom ingredient.
-			buf.writeVarInt(PACKET_MARKER);
-			buf.writeIdentifier(customIngredient.getSerializer().getIdentifier());
-			customIngredient.getSerializer().write(buf, coerceIngredient());
-		}
-	}
-
-	@Override
-	public JsonElement toJson() {
-		JsonObject json = new JsonObject();
-		json.addProperty(TYPE_KEY, customIngredient.getSerializer().getIdentifier().toString());
-		customIngredient.getSerializer().write(json, coerceIngredient());
-		return json;
-	}
-
-	@Override
-	public boolean isEmpty() {
-		// We don't want to resolve the matching stacks,
-		// as this might cause the ingredient to use outdated tags when it's done too early.
-		// So we just return false when the matching stacks haven't been resolved yet (i.e. when the field is null).
-		// TODO: this is a bit hacky, can we not do better?
-		return matchingStacks != null && matchingStacks.length == 0;
-	}
-
-	private <T> T coerceIngredient() {
-		return (T) customIngredient;
-	}
-}
