--- fabric-ingredient-api-v1/src/main/resources/fabric-ingredient-api-v1.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.ingredient",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "IngredientMixin",
-    "PacketEncoderMixin",
-    "ShapelessRecipeMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
