--- fabric-ingredient-api-v1/src/main/resources/fabric.mod.json
@@ -1,41 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-ingredient-api-v1",
-  "name": "Fabric Ingredient API (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-ingredient-api-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "mixins": [
-    "fabric-ingredient-api-v1.mixins.json"
-  ],
-  "accessWidener": "fabric-ingredient-api-v1.accesswidener",
-  "depends": {
-    "fabricloader": ">=0.14.10",
-    "fabric-networking-api-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.impl.ingredient.CustomIngredientSync"
-    ],
-    "client": [
-      "net.fabricmc.fabric.impl.ingredient.client.CustomIngredientSyncClient"
-    ]
-  },
-  "description": "Creation of new types of recipe ingredients.",
-  "custom": {
-    "fabric-api:module-lifecycle": "stable",
-    "loom:injected_interfaces": {
-      "net/minecraft/class_1856": ["net/fabricmc/fabric/api/ingredient/v1/FabricIngredient"]
-    }
-  }
-}
