--- fabric-ingredient-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,18 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-ingredient-api-v1-testmod",
-  "name": "Fabric Ingredient API (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-ingredient-api-v1":"*"
-  },
-  "entrypoints": {
-    "fabric-gametest": [
-      "net.fabricmc.fabric.test.ingredient.IngredientMatchTests",
-      "net.fabricmc.fabric.test.ingredient.SerializationTests",
-      "net.fabricmc.fabric.test.ingredient.ShapelessRecipeMatchTests"
-    ]
-  }
-}
