+++ fabric-recipe-api-v1/src/main/resources/assets/fabric-recipe-api-v1/icon.png
