+++ fabric-recipe-api-v1/src/main/resources/fabric-recipe-api-v1.accesswidener
