+++ fabric-recipe-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,18 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-recipe-api-v1-testmod",
+  "name": "Fabric Recipe API (v1) Test Mod",
+  "version": "1.0.0",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "depends": {
+    "fabric-recipe-api-v1":"*"
+  },
+  "entrypoints": {
+    "fabric-gametest": [
+      "net.fabricmc.fabric.test.recipe.ingredient.IngredientMatchTests",
+      "net.fabricmc.fabric.test.recipe.ingredient.SerializationTests",
+      "net.fabricmc.fabric.test.recipe.ingredient.ShapelessRecipeMatchTests"
+    ]
+  }
+}
