+++ gradle.properties
@@ -28,7 +28,6 @@ fabric-events-interaction-v0-version=0.4.38
 fabric-events-lifecycle-v0-version=0.2.40
 fabric-game-rule-api-v1-version=1.0.29
 fabric-gametest-api-v1-version=1.1.15
-fabric-ingredient-api-v1-version=1.0.0
 fabric-item-api-v1-version=2.1.6
 fabric-item-group-api-v1-version=2.1.9
 fabric-key-binding-api-v1-version=1.0.29
@@ -43,6 +42,7 @@ fabric-networking-api-v1-version=1.2.15
 fabric-networking-v0-version=0.3.32
 fabric-object-builder-api-v1-version=5.2.1
 fabric-particles-v1-version=1.0.19
+fabric-recipe-api-v1-version=1.0.0
 fabric-registry-sync-v0-version=2.0.2
 fabric-renderer-api-v1-version=2.1.1
 fabric-renderer-indigo-version=0.6.24
