+++ settings.gradle
@@ -26,7 +26,6 @@ include 'fabric-entity-events-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-game-rule-api-v1'
 include 'fabric-gametest-api-v1'
-include 'fabric-ingredient-api-v1'
 include 'fabric-item-api-v1'
 include 'fabric-item-group-api-v1'
 include 'fabric-key-binding-api-v1'
@@ -38,6 +37,7 @@ include 'fabric-models-v0'
 include 'fabric-networking-api-v1'
 include 'fabric-object-builder-api-v1'
 include 'fabric-particles-v1'
+include 'fabric-recipe-api-v1'
 include 'fabric-registry-sync-v0'
 include 'fabric-renderer-api-v1'
 include 'fabric-renderer-indigo'
